feat: use monitoring strategy for display and extract cmp
