+++ src/frontend/src/lib/components/modals/MonitoringDetailsModal.svelte
@@ -14,7 +14,7 @@
 	import type { JunoModalDetail, JunoModalShowMonitoringDetail } from '$lib/types/modal';
 	import { formatTCycles } from '$lib/utils/cycles.utils';
 	import { formatToRelativeTime } from '$lib/utils/date.utils';
-	import { i18nFormat } from '$lib/utils/i18n.utils';
+	import MonitoringSentence from '$lib/components/modals/MonitoringSentence.svelte';
 
 	interface Props {
 		detail: JunoModalDetail;
@@ -57,18 +57,7 @@
 								{$i18n.monitoring.auto_refill}
 							{/snippet}
 
-							<p>
-								{i18nFormat($i18n.monitoring.auto_refill_strategy, [
-									{
-										placeholder: '{0}',
-										value: formatTCycles(monitoringStrategy.BelowThreshold.min_cycles)
-									},
-									{
-										placeholder: '{1}',
-										value: formatTCycles(monitoringStrategy.BelowThreshold.fund_cycles)
-									}
-								])}
-							</p>
+							<p><MonitoringSentence {monitoringStrategy} /></p>
 						</Value>
 					{:else}
 						<MonitoringDisabled {monitoring} loading={false} />
